LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/access-from-coop-page-to-opener_coop-ro_cross-origin.https-expected.txt


FAIL access-from-coop-page-to-opener, cross-origin assert_equals: expected (string) "coop" but got (undefined) undefined
FAIL access-from-coop-page-to-opener, cross-origin + redirect assert_equals: expected (string) "coop" but got (undefined) undefined
